requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.4.9
numpy==2.2.6
packaging==25.0
pillow==12.1.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
